may have finished backend

+++ helpful_app/views/appointment_views.py
@@ -3,9 +3,11 @@ from django.http import JsonResponse
 from django.shortcuts import get_object_or_404
 from rest_framework import generics, status
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import PermissionDenied
 
 from ..models.appointment import Appointment
-from ..serializers import AppointmentSerializer
+from ..serializers import AppointmentSerializer, AppointmentReadSerializer
 
 
 class AppointmentsView(generics.ListCreateAPIView):
@@ -14,14 +16,18 @@ class AppointmentsView(generics.ListCreateAPIView):
 
     /library/appointments/
     """
+    permission_classes = (IsAuthenticated,)
     serializer_class = AppointmentSerializer
 
     def get(self, request):
-        appointments = Appointment.objects.all()
-        serializer = AppointmentSerializer(appointments, many=True)
+        '''Index request'''
+        appointments = Appointment.objects.filter(owner=request.user.id)
+        serializer = AppointmentReadSerializer(appointments, many=True)
         return Response({'appointments': serializer.data})
 
     def post(self, request):
+        '''Create request'''
+        request.data['appointment']['owner'] = request.user.id
         serializer = AppointmentSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -31,16 +37,24 @@ class AppointmentsView(generics.ListCreateAPIView):
 # /appointments/id
 
 
-class AppointmentDetailView(generics.ListCreateAPIView):
+class AppointmentDetailView(generics.RetrieveUpdateDestroyAPIView):
+    permission_classes = (IsAuthenticated,)
     serializer_class = AppointmentSerializer
 
     def get(self, request, pk):
+        '''Show request'''
         appointment = get_object_or_404(Appointment, pk=pk)
-        serializer = AppointmentSerializer(appointment)
+        if not request.user.id == appointment.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this appointment')
+        serializer = AppointmentReadSerializer(appointment)
         return Response(serializer.data)
 
     def patch(self, request, pk):
+        '''Update Request'''
         appointment = get_object_or_404(Appointment, pk=pk)
+        if not request.user.id == appointment.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this appointment')
+        request.data['appointment']['owner'] = request.user.id
         serializer = AppointmentSerializer(appointment, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -48,7 +62,10 @@ class AppointmentDetailView(generics.ListCreateAPIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
+        '''Delete Request'''
         appointment = get_object_or_404(Appointment, pk=pk)
+        if not request.user.id == appointment.owner.id:
+            raise PermissionDenied('Unauthorized, you do not own this appointment')
         appointment.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
